requirements.txt
langgraph>=1.0.0
langchain-ollama>=1.0.0
langchain-core>=1.2.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.7.4
markdown
weasyprint
